requirements.txt
# Django Core
Django==4.2.11
djangorestframework==3.14.0

# Database - PostgreSQL
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8

# API & Serialization (Comunicación entre servicios)
requests==2.31.0

# Utilities
django-filter==23.5
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
Pillow==10.2.0
<<<<<<< HEAD
qrcode[pil]==7.4.2
reportlab==4.1.0
=======
qrcode[pil]==7.4.2
>>>>>>> 9507609 (Subiendo proyecto parte frontend Marcia)
